chore: moved Select/SingleSelect from SingleSelect.tsx => Select.tsx
