+++ src/components/Select/index.ts
@@ -1,10 +1,9 @@
-export { SelectGroup, SelectRoot, SelectValue } from './src/Select';
+export { Select, SelectGroup, SelectRoot, SelectValue, SingleSelect } from './src/Select';
 export { SelectContent } from './src/SelectContent';
 export { SelectItem } from './src/SelectItem';
 export { SelectLabel } from './src/SelectLabel';
 export { SelectSeparator } from './src/SelectSeparator';
 export { SelectTrigger } from './src/SelectTrigger';
-export { Select, SingleSelect } from './src/SingleSelect';
 export type {
   SelectContentProps,
   SelectItemProps,
